feat(backend)!(frontend): Refactor balance system to support multiple years
BREAKING CHANGE: Balance API endpoints now require year parameter

Backend changes:
- Change Balance primary key from Byte (id=1) to Short (year)
- Add endpoints: GET /balance/{year}, /balance/with-totals/{year}, /balance/years
- Add PUT/DELETE operations for year-specific balance management
- Refactor BalanceService to handle year-based queries
- Update DTO constructors and mappers

Frontend changes:
- Update API endpoint configurations in settings.json files
- Add year selector dropdown in BalanceReport component
- Rewrite Balance page to fetch available years and handle year selection
- Update balance report CSS styling
- Add year filtering UI components to select balance data annually

+++ backend/huertos/src/main/java/es/huertosbellavista/backend/huertos/repository/BalanceRepository.java
@@ -3,6 +3,6 @@ package es.huertosbellavista.backend.huertos.repository;
 import es.huertosbellavista.backend.huertos.model.Balance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BalanceRepository extends JpaRepository<Balance, Byte> {
+public interface BalanceRepository extends JpaRepository<Balance, Short> {
 
 }
